requirements.txt

dj-database-url==0.5.0
Django==3.0.5
djangorestframework==3.11.0
pytz==2019.2
sqlparse==0.3.0
dj-database-url==0.5.0
gunicorn==19.9.0 
psycopg2-binary==2.7.7
urllib3==1.23
whitenoise==4.1.2